DFT View: Initialize input clip
Currently, if the source clip is missing any relevant props for outputting and they need to be assumed, vspreview will throw ANOTHER warning for DFT view. However, the colourspace is irrelevant for this plugin. Worse still, for SD content, the final size of the frame is large enough that it will instead suggest setting it to BT709, which is wrong for many DVDs.

To remedy this, we just assume properties ourselves.